requirements.txt
bidict==0.23.1
blinker==1.9.0
branca==0.8.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
eventlet==0.33.3
filelock==3.19.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
folium==0.14.0
geographiclib==2.1
geopy==2.3.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.14.1
requests==2.32.5
segment-anything==1.0
simple-websocket==1.1.0
six==1.17.0
sympy==1.14.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
gunicorn==21.2.0
setuptools==68.2.2
wheel==0.42.0
fpdf2==2.7.9
--extra-index-url https://download.pytorch.org/whl/cpu
gdown==5.2.0
